refactor(server): extract logging setup into logging.rs (main.rs was over the cap)
The log-spam fix tipped main.rs past the 700-line hard cap. Move the logging
cluster (init_logging + archive_previous_logs + chrono_timestamp + days_to_ymd,
~370 lines) into a new crates/server/src/logging.rs along the natural seam;
main.rs drops to 338 lines, logging.rs is 391. Pure move + pub(crate) on
init_logging; no behavior change.
