+++ crates/server/src/logging.rs
@@ -1,0 +1,391 @@
+//! Tracing subscriber setup for the Cimmeria server: log-file archival,
+//! the per-target file layers, the console layer, the in-memory admin-WS
+//! broadcast layer, and the optional OTLP layers. Extracted from `main.rs`
+//! to keep that file under the size cap; `main` calls [`init_logging`] once
+//! at startup and holds the returned `WorkerGuard`s for the process lifetime.
+
+use std::fs;
+use std::path::Path;
+
+use tokio::sync::broadcast;
+use tracing_appender::non_blocking::WorkerGuard;
+use tracing_subscriber::fmt;
+use tracing_subscriber::layer::SubscriberExt;
+use tracing_subscriber::util::SubscriberInitExt;
+use tracing_subscriber::EnvFilter;
+use tracing_subscriber::Layer;
+
+use cimmeria_admin_api::ws::broadcast_layer::{BroadcastLayer, LogBuffer, LogEntry};
+
+use crate::otel;
+
+// ── Logging ──────────────────────────────────────────────────────────────────
+
+/// Archive any `.log` files from a previous session into `logs/archive/<timestamp>/`.
+fn archive_previous_logs() {
+    let logs_dir = Path::new("logs");
+    if !logs_dir.exists() {
+        return;
+    }
+
+    let entries: Vec<_> = fs::read_dir(logs_dir)
+        .into_iter()
+        .flatten()
+        .filter_map(|e| e.ok())
+        .filter(|e| e.path().extension().is_some_and(|ext| ext == "log"))
+        .collect();
+
+    if entries.is_empty() {
+        return;
+    }
+
+    let ts = chrono_timestamp();
+    let archive_dir = logs_dir.join("archive").join(&ts);
+    if let Err(e) = fs::create_dir_all(&archive_dir) {
+        eprintln!(
+            "Failed to create archive directory {}: {e}",
+            archive_dir.display()
+        );
+        return;
+    }
+
+    for entry in &entries {
+        let src = entry.path();
+        let dst = archive_dir.join(entry.file_name());
+        if let Err(e) = fs::rename(&src, &dst) {
+            eprintln!("Failed to archive {}: {e}", src.display());
+        }
+    }
+
+    eprintln!(
+        "Archived {} log file(s) to {}",
+        entries.len(),
+        archive_dir.display()
+    );
+}
+
+/// Generate a filesystem-safe timestamp like `2026-03-03T14-30-22`.
+fn chrono_timestamp() -> String {
+    use std::time::SystemTime;
+
+    let now = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_secs();
+
+    // Simple UTC breakdown (no chrono dependency needed).
+    let secs_per_day: u64 = 86400;
+    let days = now / secs_per_day;
+    let day_secs = now % secs_per_day;
+    let hours = day_secs / 3600;
+    let minutes = (day_secs % 3600) / 60;
+    let seconds = day_secs % 60;
+
+    // Days since epoch to Y-M-D (simplified Gregorian).
+    let (year, month, day) = days_to_ymd(days);
+
+    format!(
+        "{:04}-{:02}-{:02}T{:02}-{:02}-{:02}",
+        year, month, day, hours, minutes, seconds
+    )
+}
+
+/// Convert days since Unix epoch to (year, month, day).
+fn days_to_ymd(days: u64) -> (u64, u64, u64) {
+    // Algorithm from Howard Hinnant's `civil_from_days`.
+    let z = days as i64 + 719468;
+    let era = if z >= 0 { z } else { z - 146096 } / 146097;
+    let doe = (z - era * 146097) as u64;
+    let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
+    let y = yoe as i64 + era * 400;
+    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
+    let mp = (5 * doy + 2) / 153;
+    let d = doy - (153 * mp + 2) / 5 + 1;
+    let m = if mp < 10 { mp + 3 } else { mp - 9 };
+    let y = if m <= 2 { y + 1 } else { y };
+    (y as u64, m, d)
+}
+
+/// Initialise a layered tracing subscriber.
+///
+/// Returns `WorkerGuard`s that **must** be held alive in `main()` — dropping
+/// them flushes and closes the log files.
+///
+/// Layers:
+/// - **Console** (stdout): coloured, human-readable, level from `RUST_LOG` (default `info`).
+/// - **`logs/server.log`**: JSON, all modules at `info` (high-level milestones only).
+/// - **`logs/auth.log`**: auth module at `trace`.
+/// - **`logs/base.log`**: connection lifecycle (service, connect_loop, login, tick_sync, helpers) at `trace`.
+/// - **`logs/world_entry.log`**: player world entry, DB queries, appearance at `trace`.
+/// - **`logs/character.log`**: char list, creation, visuals, PAK serving at `trace`.
+/// - **`logs/protocol.log`**: wire-level packet building, encryption at `trace`.
+/// - **`logs/aoi.log`**: AoI tick, entity create/destroy, spatial grid at `trace`.
+/// - **`logs/combat.log`**: damage, abilities, NPC AI, death/respawn at `trace`.
+/// - **`logs/content.log`**: content chain triggers, conditions, actions at `trace`.
+/// - **`logs/missions.log`**: mission accept/complete/abandon, objectives at `trace`.
+/// - **`logs/interactions.log`**: dialogs, vendors, trainers, loot, chat, mail at `trace`.
+/// - **`logs/spawner.log`**: NPC spawning, stargate travel, ring transport at `trace`.
+/// - **`logs/dispatch.log`**: method routing (cell + base method dispatch) at `trace`.
+/// - **WebSocket broadcast**: `debug` for admin panel streaming.
+/// - **Cosmos DB** (optional): `debug` minus per-packet noise.
+///
+/// Correlation: most tracing calls include `entity_id`, `player_id`, or
+/// `witness_id` as structured fields.  To follow a player across files:
+/// `grep entity_id=42 logs/*.log`.
+pub(crate) fn init_logging(
+    log_tx: broadcast::Sender<LogEntry>,
+    log_buffer: LogBuffer,
+    otel_trace_layer: Option<otel::OtelTraceLayer>,
+    otel_log_layer: Option<otel::OtelLogLayer>,
+    otel_network_log_layer: Option<otel::OtelLogLayer>,
+) -> Vec<WorkerGuard> {
+    // Move previous session's logs into archive/.
+    archive_previous_logs();
+
+    // Ensure logs/ directory exists.
+    let _ = fs::create_dir_all("logs");
+
+    let mut guards = Vec::new();
+
+    type BoxLayer = Box<dyn tracing_subscriber::Layer<tracing_subscriber::Registry> + Send + Sync>;
+
+    // Helper: create a plain-text, non-ANSI log file layer with a specific filter.
+    // Returns a boxed layer to avoid deeply nested generic types that cause
+    // the Rust type-checker to consume 50+ GB of RAM.
+    macro_rules! log_layer {
+        ($filename:expr, $filter:expr) => {{
+            let file = tracing_appender::rolling::never("logs", $filename);
+            let (writer, guard) = tracing_appender::non_blocking(file);
+            guards.push(guard);
+            Box::new(
+                fmt::layer()
+                    .with_writer(writer)
+                    .with_ansi(false)
+                    .with_target(true)
+                    .with_filter(EnvFilter::new($filter)),
+            ) as BoxLayer
+        }};
+    }
+
+    // ── Collect all layers into a Vec<BoxLayer> ─────────────────────────
+    let mut layers: Vec<BoxLayer> = Vec::new();
+
+    // ── Console (stdout, coloured, RUST_LOG or info) ─────────────────────
+    // Default to `info` but mute the per-packet wire firehose — `mercury.packet`
+    // (one INFO event per UDP datagram), the `wire.in`/`wire.out` method-call
+    // stream, and retransmit noise. These stay at full fidelity in the file
+    // layers (protocol.log) and OTLP; they just don't belong on an operator's
+    // console. `RUST_LOG`, when set, overrides this entirely.
+    const CONSOLE_DEFAULT: &str =
+        "info,mercury.packet=warn,wire.in=warn,wire.out=warn,mercury.retransmit=warn";
+    let console_filter =
+        EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new(CONSOLE_DEFAULT));
+    layers.push(Box::new(fmt::layer().with_filter(console_filter)));
+
+    // ── server.log (JSON, all modules, info) ─────────────────────────────
+    let server_file = tracing_appender::rolling::never("logs", "server.log");
+    let (server_writer, guard) = tracing_appender::non_blocking(server_file);
+    guards.push(guard);
+    layers.push(Box::new(
+        fmt::layer()
+            .json()
+            .with_writer(server_writer)
+            .with_target(true)
+            .with_filter(EnvFilter::new("info")),
+    ));
+
+    // ── Per-system log files ─────────────────────────────────────────────
+    layers.push(log_layer!("auth.log", "off,cimmeria_services::auth=trace"));
+
+    layers.push(log_layer!(
+        "base.log",
+        "off,\
+         cimmeria_services::base::service=trace,\
+         cimmeria_services::base::connect_loop=trace,\
+         cimmeria_services::base::login=trace,\
+         cimmeria_services::base::tick_sync=trace,\
+         cimmeria_services::base::helpers=trace"
+    ));
+
+    layers.push(log_layer!(
+        "world_entry.log",
+        "off,\
+         cimmeria_services::base::world_entry=trace,\
+         cimmeria_services::base::world_entry_player=trace,\
+         cimmeria_services::base::world_entry_appearance=trace"
+    ));
+
+    layers.push(log_layer!(
+        "character.log",
+        "off,\
+         cimmeria_services::base::character=trace,\
+         cimmeria_services::base::character_create=trace,\
+         cimmeria_services::base::chardef=trace,\
+         cimmeria_services::base::cooked_data=trace,\
+         cimmeria_services::base::resources=trace"
+    ));
+
+    layers.push(log_layer!(
+        "protocol.log",
+        "off,\
+         cimmeria_services::mercury=trace,\
+         cimmeria_mercury=trace,\
+         mercury.packet=info,\
+         wire.in=info,wire.out=info,\
+         mercury.retransmit=info"
+    ));
+
+    layers.push(log_layer!(
+        "aoi.log",
+        "off,\
+         cimmeria_services::cell::service=trace,\
+         cimmeria_services::cell::space_manager=trace"
+    ));
+
+    layers.push(log_layer!(
+        "combat.log",
+        "off,\
+         cimmeria_services::cell::combat=trace,\
+         cimmeria_services::cell::abilities=trace"
+    ));
+
+    layers.push(log_layer!(
+        "content.log",
+        "off,cimmeria_services::cell::content=trace"
+    ));
+
+    layers.push(log_layer!(
+        "missions.log",
+        "off,cimmeria_services::cell::missions=trace"
+    ));
+
+    layers.push(log_layer!(
+        "interactions.log",
+        "off,\
+         cimmeria_services::cell::interactions=trace,\
+         cimmeria_services::cell::chat=trace,\
+         cimmeria_services::cell::mail=trace"
+    ));
+
+    layers.push(log_layer!(
+        "spawner.log",
+        "off,\
+         cimmeria_services::cell::spawner=trace,\
+         cimmeria_services::cell::gate_travel=trace,\
+         cimmeria_services::cell::ring_transport=trace"
+    ));
+
+    layers.push(log_layer!(
+        "dispatch.log",
+        "off,\
+         cimmeria_services::cell::dispatch=trace,\
+         cimmeria_services::base::dispatch=trace"
+    ));
+
+    // ── WebSocket broadcast (debug+, all modules) ─────────────────────
+    layers.push(Box::new(
+        BroadcastLayer::new(log_tx, log_buffer).with_filter(EnvFilter::new(
+            "debug,tungstenite=info,tokio_tungstenite=info,hyper=info",
+        )),
+    ));
+
+    // ── Discord notifications (optional — config-gated) ──────────────
+    // The layer harvests `warn!`/`error!` events with structured fields
+    // and posts them to the configured Discord channels. Disabled events
+    // and missing channels short-circuit before the queue, so the layer
+    // is cheap when Discord is off.
+    if let Some(rt) = cimmeria_discord::global() {
+        layers.push(Box::new(
+            cimmeria_discord::DiscordLayer::new(rt.handle.clone(), rt.config.handle())
+                // Layer applies its own per-event toggle gating inside
+                // on_event — no env-filter needed here. Keep the
+                // env-filter wide so we don't pre-filter out events the
+                // user might want to enable at runtime via toggle.
+                .with_filter(EnvFilter::new("warn")),
+        ));
+    }
+
+    // ── OpenTelemetry → SigNoz (optional) ─────────────────────────────
+    // Two layers come through together when OTLP is enabled:
+    //
+    //   * `trace_layer` — captures `tracing::span!` spans (and events
+    //     fired *inside* them) as OpenTelemetry spans. Shows up in the
+    //     SigNoz Traces view.
+    //   * `log_layer`   — captures every `tracing::*` event as an OTLP
+    //     log record, including root-level events that have no parent
+    //     span. The Mercury packet stream lives here. Shows up in the
+    //     SigNoz Logs view.
+    //
+    // Both layers share the same env-filter so the operator only has to
+    // tune visibility once. Filters ship `debug+` from our crates and
+    // drop the chatty HTTP middleware noise; `mercury.packet` is kept
+    // at info — it's the load-bearing analytical surface, sampling
+    // would defeat the purpose.
+    let otel_filter = "info,\
+                cimmeria_services=debug,\
+                cimmeria_mercury=debug,\
+                mercury.packet=info,\
+                mercury.retransmit=info,\
+                mercury.backpressure=warn,\
+                wire.in=info,wire.out=info,\
+                aoi.entity_enter=debug,aoi.entity_leave=debug,\
+                movement.npc=debug,movement.player=debug,\
+                npc_ai=debug,\
+                threat=info,\
+                auth=info,\
+                world_entry=info,\
+                vendor=info,mail=info,progression=info,inventory=info,mission=info,\
+                sqlx::query=debug,\
+                tungstenite=off,tokio_tungstenite=off,hyper=off,\
+                h2=off,tower=off,tonic=off,reqwest=off,opentelemetry=off";
+
+    if let Some(layer) = otel_trace_layer {
+        layers.push(Box::new(layer.with_filter(EnvFilter::new(otel_filter))));
+    }
+    // The log signal splits across TWO OTLP providers (cimmeria-server +
+    // cimmeria-network — see `otel::init`). Per-layer FilterFn routes:
+    //
+    // - `cimmeria-server` layer: receive everything EXCEPT TRACE/DEBUG/
+    //   INFO from network-noise scopes. WARN+ from network-noise scopes
+    //   still goes here so elevated severity surfaces in the operator's
+    //   primary view without dual-querying.
+    // - `cimmeria-network` layer: receive ONLY TRACE/DEBUG/INFO from
+    //   network-noise scopes. WARN+ is suppressed (it's already in the
+    //   server index).
+    //
+    // Composition: each branch is `EnvFilter::new(otel_filter)` AND a
+    // `FilterFn` doing the noise routing. Together they cover all events
+    // with no overlap (a single event lands in exactly one log index).
+    //
+    // We use `tracing_subscriber::filter::FilterExt::and` to combine the
+    // two filters, and `Box::new` at the end because the layer push
+    // signature wants a homogeneous trait object.
+    if let Some(layer) = otel_log_layer {
+        use tracing_subscriber::filter::{filter_fn, FilterExt};
+        let routing = filter_fn(|meta| {
+            // Server index: non-noise OR severity is WARN/ERROR.
+            !otel::is_network_noise_target(meta.target()) || *meta.level() <= tracing::Level::WARN
+        });
+        layers.push(Box::new(
+            layer.with_filter(EnvFilter::new(otel_filter).and(routing)),
+        ));
+    }
+    if let Some(layer) = otel_network_log_layer {
+        use tracing_subscriber::filter::{filter_fn, FilterExt};
+        let routing = filter_fn(|meta| {
+            // Network index: noise scopes only, at non-elevated severity.
+            // `level <= Level::WARN` is "WARN or more severe" (numerically
+            // smaller); the negation here means "less severe than WARN"
+            // i.e. INFO/DEBUG/TRACE.
+            otel::is_network_noise_target(meta.target()) && *meta.level() > tracing::Level::WARN
+        });
+        layers.push(Box::new(
+            layer.with_filter(EnvFilter::new(otel_filter).and(routing)),
+        ));
+    }
+
+    // Assemble the subscriber — one `.with()` call on the whole Vec.
+    tracing_subscriber::registry().with(layers).init();
+
+    guards
+}
